# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProfileUpsert": {
        "additionalProperties": false,
        "description": "PUT body for one profile. ``extra=\"forbid\"`` is the AC-A.10\ntripwire: a raw ``token`` member is an unknown field \u2192 422.",
        "properties": {
          "caps": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-tool per-session caps (tool -> max calls), superseding the built-in 3-search/4-chunk constants.",
            "title": "Caps"
          },
          "enabled": {
            "default": true,
            "title": "Enabled",
            "type": "boolean"
          },
          "notebooks": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Notebook-slug allowlist; null = unscoped.",
            "title": "Notebooks"
          },
          "token_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "SHA-256 hex digest of the bearer token. NEVER the token value. Omit for the token-less default profile.",
            "title": "Token Sha256"
          },
          "tools": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Tool allowlist; null = all tools.",
            "title": "Tools"
          }
        },
        "title": "ProfileUpsert",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "JSON-only operator surface of the arXMCP server: /api/v1 notebook management aliases, health/readiness/status probes, and the /bridge/contracts handshake. Generated OFFLINE by tools/dump_openapi.py (IF-1); the runtime openapi_url is disabled by design (Threat 4).",
    "title": "arXMCP operator API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/capabilities/profiles/{name}": {
      "delete": {
        "description": "Remove one profile. Deleting an operator-defined ``default``\nrestores the synthesized permissive default. 404 when absent.",
        "operationId": "delete_profile_api_v1_capabilities_profiles__name__delete",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Profile",
        "tags": [
          "capabilities"
        ]
      },
      "put": {
        "description": "Create or replace one profile; effective immediately (cache\ninvalidated inline).",
        "operationId": "upsert_profile_api_v1_capabilities_profiles__name__put",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upsert Profile Api V1 Capabilities Profiles  Name  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Profile",
        "tags": [
          "capabilities"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "http://127.0.0.1:7733"
    }
  ]
}